requirements.txt
tensorflow
numpy
pillow
opencv--python
matplotlib
serial